+++ src/components/cards/HomeProjectCard.tsx
@@ -1,7 +1,7 @@
-"use client";
-import Link from "next/link";
-import { getProjectSlug } from "@/utils";
-import type { HomeProjectCardProps } from "@/types/components";
+'use client';
+import Link from 'next/link';
+import { getProjectSlug } from '@/utils';
+import type { HomeProjectCardProps } from '@/types/components';
 
 export function HomeProjectCard({
   name,
@@ -10,9 +10,9 @@ export function HomeProjectCard({
   link,
 }: HomeProjectCardProps) {
   const handleClick = (e: React.MouseEvent) => {
-    if (link && !link.startsWith("/")) {
+    if (link && !link.startsWith('/')) {
       e.preventDefault();
-      window.open(link, "_blank");
+      window.open(link, '_blank');
     }
   };
 
@@ -22,7 +22,7 @@ export function HomeProjectCard({
 
   const cardContent = (
     <div
-      onClick={link && !link.startsWith("/") ? handleClick : undefined}
+      onClick={link && !link.startsWith('/') ? handleClick : undefined}
       className={`border-2 border-slate-800/70 bg-white font-extralight h-24 lg:h-32 rounded-lg flex items-center text-2xl sm:text-3xl hover:cursor-pointer ${color.text} transition-all duration-300 p-10 hover:-translate-y-0.5 hover:shadow-md flex justify-start sm:justify-center`}
     >
       {name}
@@ -33,8 +33,8 @@ export function HomeProjectCard({
     return <Link href={projectPath}>{cardContent}</Link>;
   }
 
-  if (link?.startsWith("/")) {
-    return <Link href={link}>{cardContent}</Link>;
+  if (link?.startsWith('/')) {
+    return <a href={link}>{cardContent}</a>;
   }
 
   return <div>{cardContent}</div>;